Adds visibility to container element
Adding `visibility: hidden` by default and `visibility: visible` on
active will prevent clickthrough of the links when the tooltip is
not active. Previously, the tooltip was clickable in the
top center of the page on load, with a z-index that would make some
content unclickable if it happened in the same position. This treats the
element as if it isn't present, without reflowing the DOM the same way
that `display: none;` does.